Move put with optional last updated to BaseModel
